Moved all bid related files to a separate file under route /bids
