+++ bio-bid/src/components/bids/styles/bids.css
