+++ bio-bid/src/components/bids/styles/dash.css
