+++ bio-bid/src/components/bids/styles/global.css
