+++ bio-bid/src/components/bids/styles/main.css
@@ -1,10 +1,3 @@
-.App {
-  text-align: center;
-  width: 100%;
-  height:100%;
-  margin: 0 auto;
-}
-
 .shadow-active {
 
     z-index: 1;
