+++ bio-bid/src/components/bids/styles/open-sidebar.css
