+++ bio-bid/src/components/bids/styles/styled-componets.js
@@ -9,5 +9,5 @@ export const WorkSpace =styled.div
 `
 display:flex;
 min-height: 100%;
-height: 100%;
+height: 100vh;
 `;
