+++ bio-bid/src/components/company-profile/Form/Form.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default () => {
+    return (
+        <div>
+            <h1>Form</h1>
+        </div>
+    );
+}
