+++ bio-bid/src/components/company-profile/Form/styles.js
@@ -1,0 +1,2 @@
+import styled from 'styled-components';
+import theme from '../../../theme';
